+++ App/src/app_crc.c
@@ -24,8 +24,8 @@ uint8_t app_CalcCRC8(uint8_t *ptr, uint8_t len) {
 	return crc;
 }
 
-uint8_t app_CalcCRC8_cycle(uint8_t *ptr, uint8_t len, uint8_t pos,
-		uint8_t bufLen) {
+uint8_t app_CalcCRC8_cycle(uint8_t *ptr, uint16_t len, uint16_t pos,
+		uint16_t bufLen) {
 	crc = 0;
 	while (len--) {
 		crc ^= *ptr++;
